apps/api/pyproject.toml
[project]
name = "backpacking-assistant-api"
version = "0.1.0"
description = "FastAPI backend with LangGraph agentic architecture for backpacking trip planning"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "pydantic==2.10.3",
    "pydantic-settings==2.6.1",
    "python-dotenv==1.0.1",
    "langgraph==0.2.59",
    "langchain-core==0.3.27",
    "langchain-google-genai==2.0.7",
    "google-generativeai==0.8.3",
    "httpx==0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "black==24.10.0",
    "ruff==0.8.4",
]
full = [
    "supabase==2.11.0",
    "python-socketio==5.12.0",
    "redis==5.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
